Duplicate flags allowed from different plugins
All flags are saved in the database prefixed with the plugin's
name ase returned by Plugin.getName(). If the flag name is unique
across all plugins, then the user only sees the short name. If
there's a duplicate flag registered from two different plugins (or
a plugin and a builtin flag) or a plugin is missing, then the user
will see the full flag name with the prefixed plugin name included.